+++ concrete_datastore/concrete/middleware.py
@@ -5,9 +5,9 @@ from __future__ import (
     print_function,
     unicode_literals,
 )
-
+import pendulum
 from django_otp.middleware import OTPMiddleware
-
+from concrete_datastore.api.v1.datetime import format_datetime
 from concrete_datastore.concrete.models import EmailDevice
 
 
@@ -20,3 +20,16 @@ class OTPCustomMiddleware(OTPMiddleware):
         device = EmailDevice.from_persistent_id(persistent_id)
 
         return device
+
+
+class DateTimeLoggerMiddleware:
+    def __init__(self, get_response):
+        self.get_response = get_response
+
+    def __call__(self, request):
+        date_received = pendulum.now('utc')
+        response = self.get_response(request)
+        date_sent = pendulum.now('utc')
+        response['DateTime-Received'] = format_datetime(date_received)
+        response['DateTime-Sent'] = format_datetime(date_sent)
+        return response